fix: 料理をつくるボタンを削除、アプリをインストールしてから１回だけポップアップが表示されるように変更
