+++ lib/ui/screens/cooking_screen.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:caul/status/popup.state.dart';
 import 'package:caul/providers/popup_provider.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 final pageControllerProvider = Provider((ref) => PageController());
 
@@ -16,70 +17,9 @@ class CookingScreen extends ConsumerWidget {
     final popupNotifier = ref.watch(popupProvider.notifier);
     final pageController = popupNotifier.pageController;
 
-    Future<void> _showPopup(BuildContext context, WidgetRef ref,
-        PageController pageController) async {
-      await showDialog(
-        context: context,
-        barrierDismissible: true,
-        builder: (BuildContext context) {
-          return PageView.builder(
-            controller: pageController,
-            itemCount: 3,
-            onPageChanged: (index) {
-              ref.read(popupProvider.notifier).updateCurrentPage(index + 1);
-            },
-            itemBuilder: (context, index) {
-              String description;
-              String buttonLabel;
-              VoidCallback onPressed;
-
-              switch (index) {
-                case 0:
-                  description = "これはページ1です。";
-                  buttonLabel = "次へ";
-                  onPressed = () => pageController.nextPage(
-                      duration: const Duration(milliseconds: 300),
-                      curve: Curves.easeInOut);
-                  break;
-                case 1:
-                  description = "これはページ2です。";
-                  buttonLabel = "次へ";
-                  onPressed = () => pageController.nextPage(
-                      duration: const Duration(milliseconds: 300),
-                      curve: Curves.easeInOut);
-                  break;
-                case 2:
-                  description = "これはページ3です。";
-                  buttonLabel = "料理を作る！";
-                  onPressed = () => Navigator.of(context).pop();
-                  break;
-                default:
-                  throw Exception("Invalid page index");
-              }
-
-              return _PopupPage(
-                description: description,
-                buttonLabel: buttonLabel,
-                onPressed: onPressed,
-              );
-            },
-          );
-        },
-      );
-      WidgetsBinding.instance?.addPostFrameCallback((_) {
-        _showPopup(context, ref, pageController);
-      });
-    }
-
     return Scaffold(
       body: Container(
         color: Colors.orange[400],
-        child: Center(
-          child: ElevatedButton(
-            onPressed: () => _showPopup(context, ref, pageController),
-            child: const Text('料理ページを開く'),
-          ),
-        ),
       ),
     );
   }
@@ -98,31 +38,5 @@ class PopupNotifier extends StateNotifier<PopupState> {
   void dispose() {
     pageController.dispose();
     super.dispose();
-  }
-}
-
-class _PopupPage extends StatelessWidget {
-  final String description;
-  final String buttonLabel;
-  final VoidCallback onPressed;
-
-  const _PopupPage({
-    Key? key,
-    required this.description,
-    required this.buttonLabel,
-    required this.onPressed,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      content: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text(description),
-          ElevatedButton(onPressed: onPressed, child: Text(buttonLabel)),
-        ],
-      ),
-    );
   }
 }
