+++ pubspec.yaml
@@ -18,6 +18,7 @@ dependencies:
   font_awesome_flutter: ^10.5.0
   google_sign_in: ^6.1.5
   flutter_riverpod: ^2.4.0
+  shared_preferences: ^2.2.1
 
 dev_dependencies:
   flutter_test:
